% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_deconvolve.R
\name{cell_deconvolve_trace}
\alias{cell_deconvolve_trace}
\title{Calculate the GEP of each cell types in bulk samples}
\usage{
cell_deconvolve_trace(
  O,
  theta,
  R,
  alpha = 0.5,
  beta = 5,
  gamma = 1,
  epsilon = 0.001,
  max.iter = 100,
  verbose = FALSE
)
}
\arguments{
\item{verbose}{}
}
\value{

}
\description{
Calculate the GEP of each cell types in bulk samples
}
