% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_model.R
\name{clean_model}
\alias{clean_model}
\title{Clean the trained models for ENIGMA objects}
\usage{
clean_model(object, model_list)
}
\arguments{
\item{object}{ENIGMA object}

\item{model_list}{input the names (list) of models}
}
\value{
ENIGMA object
}
\description{
clean the unwanted models in ENIGMA objects
}
\examples{
\dontrun{
egm = clean_model(egm,model_list = c(1,2)) # preserve the 1st and 2st models
egm = clean_model(egm,model_list = c("Wang")) # preserve the model which named as "Wang"
egm = clean_model(egm,NULL) # clean all saved model
}


}
