% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ENIGMA_class.R
\name{create_ENIGMA}
\alias{create_ENIGMA}
\title{Create a new ENIGMA object}
\usage{
create_ENIGMA(
  bulk,
  ref,
  ref_type = c("single_cell", "sort", "aggre"),
  meta_bulk = NULL,
  meta_ref = NULL,
  assay = NULL
)
}
\arguments{
\item{bulk}{raw bulk data matrix}

\item{ref}{raw reference data, could be single cell (matrix, SingleCellExperiment object or Seurat object) or FACS bulk}

\item{ref_type}{Determine the reference type. Should be either "single_cell" or "sort".}

\item{meta_bulk}{metadata of bulk matrix}

\item{meta_ref}{metadata of reference}
}
\value{
an ENIGMA initial object
}
\description{
Create a new ENIGMA object
}
