% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ENIGMA_class.R
\docType{class}
\name{ENIGMA-class}
\alias{ENIGMA-class}
\alias{ENIGMA}
\title{The ENIGMA Class}
\description{
The ENIGMA Class
}
\section{Slots}{

\describe{
\item{\code{raw_input}}{raw bulk and reference data}

\item{\code{ref_type}}{the reference type}

\item{\code{bulk}}{bulk matrix after batch correction}

\item{\code{ref}}{reference matrix after batch correction}

\item{\code{result_cell_proportion}}{the proportion of each cell types in bulk samples}

\item{\code{result_CSE}}{inferred cell type-specific expression (CSE)}

\item{\code{result_CSE_normalized}}{inferred (normalized) cell type-specific expression (CSE)}
}}

