% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ENIGMA-package.R
\docType{package}
\name{ENIGMA-package}
\alias{ENIGMA-package}
\alias{ENIGMA}
\title{The 'ENIGMA' package.}
\description{
An enterotype-like unigram class model for identifying microbial associations.
}
\references{
Stan Development Team (2018). RStan: the R interface to Stan. R package version 2.17.3. http://mc-stan.org
}
