% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enigma_datasets.R
\name{enigma_datasets}
\alias{enigma_datasets}
\title{Datasets search}
\usage{
enigma_datasets(query = NULL, parent_collection_id = NULL, limit = 500,
  offset = NULL, sort = NULL, filter = NULL, key = NULL, ...)
}
\arguments{
\item{query}{(character) Query string to return only rows that contain 
specific information (must set row_limit > 0). You cannot combine this 
with the filter parameter.}

\item{parent_collection_id}{(character) Returns only datasets whose 
immediate parent collection ID matches the one specified here.}

\item{limit}{(numeric) Number of rows of the dataset to return. Default 
(and max): 500}

\item{offset}{(numeric) Number of rows to skip. Default: 0}

\item{sort}{(character) List of dataset attributes indicating how to 
sort the datasets (currently display_name is the only attribute supported 
by this parameter). For example, sort=display_name sorts the datasets by 
display name. Since display_name is the default sort attribute, this 
parameter has no effect. You cannot combine this with the query parameter.}

\item{filter}{(numeric) Returns only datasets that match the specified 
attribute (currently display_name is the only attribute supported by 
this parameter). For example, filter=display_name=Companies returns 
only datasets named "Companies"; filter=display_name<G returns only 
datasets with names with first letter less than 'G'. You cannot combine 
this with the query parameter.}

\item{key}{(character) Required. An Enigma API key. Supply in the function 
call, or store in your \code{.Renviron} file like 
\code{ENIGMA_KEY=your key)}, or in your \code{.Rprofile} file as 
\code{options(enigmaKey = "<your key>")}, Obtain an API key by creating 
an account with Enigma at \url{http://enigma.io}, then obtain an API key 
from your account page.}

\item{...}{Named curl options passed on to \code{\link[crul]{HttpClient}}}
}
\value{
A tibble (data.frame) with columns (some contain nested lists or 
data.frame's):
\itemize{
 \item ancestors
 \item citation
 \item created_at
 \item data_updated_at
 \item description
 \item description_short
 \item display_name
 \item editable
 \item highlights
 \item id
 \item modified_at
 \item published
 \item schema_updated_at
 \item score
 \item current_snapshot
 \item parent_collection
}
}
\description{
Datasets search
}
\examples{
\dontrun{
res <- enigma_datasets(query = 'google', limit = 10)
res$ancestors
res$citation
res$data_updated_at
}
}
\references{
\url{https://app.enigma.io/api#data}
}
