% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enigma_dataset.R
\name{enigma_ds}
\alias{enigma_ds}
\title{Fetch dataset by ID and filter rows with query}
\usage{
enigma_ds(id, query = NULL, limit = 500, offset = NULL, sort = NULL,
  row_sort = NULL, key = NULL, ...)
}
\arguments{
\item{id}{(character) dataset ID. required.}

\item{query}{(character) Query string to return only rows that contain 
specific information (must set row_limit > 0).}

\item{limit}{(numeric) Number of rows of the dataset to return. Default 
(and max): 500}

\item{offset}{(numeric) Number of rows to skip. Default: 0}

\item{sort}{(character) List of dataset attributes indicating how to 
sort the datasets (currently display_name is the only attribute supported 
by this parameter). For example, sort=display_name sorts the datasets by 
display name. Since display_name is the default sort attribute, this 
parameter has no effect. You cannot combine this with the query parameter.}

\item{row_sort}{Specifies the field used to sort the records within the 
dataset. If you specify row_offset as well, the records are sorted first 
and then the offset is applied. Prepend the field name with a minus 
sign (-) to specify descending order (defaults to ascending).}

\item{key}{(character) Required. An Enigma API key. Supply in the function 
call, or store in your \code{.Renviron} file like 
\code{ENIGMA_KEY=your key)}, or in your \code{.Rprofile} file as 
\code{options(enigmaKey = "<your key>")}, Obtain an API key by creating 
an account with Enigma at \url{http://enigma.io}, then obtain an API key 
from your account page.}

\item{...}{Named curl options passed on to \code{\link[crul]{HttpClient}}}
}
\value{
A tibble (data.frame) with columns (some contain nested lists or 
data.frame's):
\itemize{
 \item ancestors
 \item citation
 \item created_at
 \item current_snapshot
 \item data_updated_at
 \item description
 \item description_short
 \item display_name
 \item editable
 \item id
 \item modified_at
 \item parent_collection
 \item published
 \item schema_updated_at
}
}
\description{
Fetch dataset by ID and filter rows with query
}
\examples{
\dontrun{
enigma_ds(id = 'a3164cd5-82e9-424b-8d75-efdd624141b1')
}
}
\references{
\url{https://app.enigma.io/api#data}
}
