% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enigma_dataset_snapshots.R
\name{enigma_ds_snapshots}
\alias{enigma_ds_snapshots}
\title{Fetch dataset by ID and filter rows with query}
\usage{
enigma_ds_snapshots(id, snapshot_id = NULL, query = NULL, limit = 500,
  offset = NULL, sort = NULL, row_sort = NULL, highlight = NULL,
  key = NULL, ...)
}
\arguments{
\item{id}{(character) dataset ID. required.}

\item{snapshot_id}{(character) snapshot ID. optional}

\item{query}{(character) Query string to return only rows that contain 
specific information (must set row_limit > 0).}

\item{limit}{(numeric) Number of rows of the dataset to return. Default 
(and max): 500}

\item{offset}{(numeric) Number of rows to skip. Default: 0}

\item{sort}{(character) List of dataset attributes indicating how to 
sort the datasets (currently display_name is the only attribute supported 
by this parameter). For example, sort=display_name sorts the datasets by 
display name. Since display_name is the default sort attribute, this 
parameter has no effect. You cannot combine this with the query parameter.}

\item{row_sort}{Specifies the field used to sort the records. If you 
specify row_offset as well, the records are sorted first and then the 
offset is applied. Prepend the field name with a minus sign (-) to 
specify descending order (defaults to ascending).}

\item{highlight}{(logical) When `TRUE`, for rows returned based on a query 
string, the endpoint sets the returned highlights attribute with positional 
information indicating where instances of the specified string occur. 
Default: `FALSE`}

\item{key}{(character) Required. An Enigma API key. Supply in the function 
call, or store in your \code{.Renviron} file like 
\code{ENIGMA_KEY=your key)}, or in your \code{.Rprofile} file as 
\code{options(enigmaKey = "<your key>")}, Obtain an API key by creating 
an account with Enigma at \url{http://enigma.io}, then obtain an API key 
from your account page.}

\item{...}{Named curl options passed on to \code{\link[crul]{HttpClient}}}
}
\description{
Fetch dataset by ID and filter rows with query
}
\examples{
\dontrun{
enigma_ds_snapshots(id = 'a3164cd5-82e9-424b-8d75-efdd624141b1')
# snapshots
enigma_ds_snapshots(id = 'a3164cd5-82e9-424b-8d75-efdd624141b1', 
   snapshot_id = "6cc93c26-a99d-4d2a-b57e-b2806ff1f45e")
}
}
