% Generated by roxygen2 (4.0.2): do not edit by hand
\name{enigma_fetch}
\alias{enigma_fetch}
\alias{enigma_read}
\title{Download a gzipped csv file of a dataset.}
\usage{
enigma_fetch(dataset = NULL, path = NULL, overwrite = TRUE, key = NULL,
  ...)

enigma_read(input)
}
\arguments{
\item{dataset}{Dataset name. Required.}

\item{path}{File name and path of output zip file. Defaults to write a zip file to your home
directory with name of the dataset, and file extension \code{.csv.gz}.}

\item{overwrite}{Will only overwrite existing path if TRUE.}

\item{key}{(character) An Enigma API key. Supply in the function call, or store in your
\code{.Rprofile} file, or do \code{options(enigmaKey = "<your key>")}. Required.}

\item{...}{Named options passed on to \code{\link[httr]{GET}}}

\item{input}{The output from \code{enigma_fetch} or a path to a file downloaded from Enigma.io}
}
\description{
Download a gzipped csv file of a dataset.
}
\examples{
\dontrun{
res <- enigma_fetch(dataset='com.crunchbase.info.companies.acquisition')
enigma_read(res)

# Piping workflow
library('dplyr')
enigma_fetch(dataset='com.crunchbase.info.companies.acquisition') \%>\%
   enigma_read \%>\%
   glimpse

# Curl debugging
library('httr')
enigma_fetch(dataset='com.crunchbase.info.companies.acquisition', config=verbose())
}
}

