% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ENIGMA_L2_norm.R
\name{ENIGMA_L2_max_norm}
\alias{ENIGMA_L2_max_norm}
\title{ENIGMA L2 max norm version}
\usage{
ENIGMA_L2_max_norm(
  object,
  alpha = 0.5,
  tao_k = 0.01,
  beta = 8000,
  epsilon = 0.001,
  max.iter = 1000,
  verbose = TRUE,
  infer = FALSE,
  pos = TRUE
)
}
\arguments{
\item{object}{ENIGMA object}

\item{alpha}{ENIGMA is a multi-objective optimization problem involve two object function, the distance function between observed bulk RNA-seq and reconstitute RNA-seq generated by weighted combination of CSE, and the distance function beween average CSE expression and cell type reference matrix. The alpha is used to determine weights of these two objects. If the alpha gets larger, the optimization attach greater importance on the the first object. Default: 0.5}

\item{tao_k}{The step size of each round of gradient decent. Default: 0.01}

\item{beta}{The regularization parameter to penalize the weight (deconvoluted expression) matrices from being too large. Default: 8000}

\item{epsilon}{Determine the stop condition in CSE updating. Default: 0.001}

\item{max.iter}{The maximum number of iterations. Default: 1000}

\item{verbose}{Whether return the information after each step of processing. Default: TRUE}

\item{infer}{Whether return the re-estimated cell type fractions. Default: FALSE}

\item{pos}{Whether set all entries in CSE is positive. Default: TRUE}
}
\description{
ENIGMA L2 max norm version
}
