% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ENIGMA_trace_norm.R
\name{ENIGMA_trace_norm}
\alias{ENIGMA_trace_norm}
\title{ENIGMA trace norm version}
\usage{
ENIGMA_trace_norm(
  object,
  alpha = 0.5,
  beta = 100,
  gamma = 1,
  epsilon = 1e-04,
  max.iter = 100,
  verbose = TRUE
)
}
\arguments{
\item{object}{ENIGMA object}

\item{alpha}{ENIGMA is a multi-objective optimization problem involve two object function,
the distance function between observed bulk RNA-seq and reconstitute RNA-seq
generated by weighted combination of CSE, and the distance function beween
average CSE expression and cell type reference matrix. The alpha is used to
determine weights of these two objects. If the alpha gets larger,
the optimization attach greater importance on the the first object.
Default: 0.5}

\item{beta}{This parameter is used to control the latent dimension of each CSE,
if this parameter gets larger, than the latent dimension of each CSE is smaller
(lower trace norm value), which means that each sample is more similar with
each others. The user need to tune this parameter based on the range of
the singular value of the bulk RNA-seq matrix.
Default: 100}

\item{gamma}{This parameter is used to control the distance between CSE (X) and auxiliary variable (Y). Default: 1}

\item{epsilon}{In trace norm based ENIGMA, the epsilon is not necessarily choose
a extremly small value, the number of iteration would influence
the latent dimensions of CSE, as each step is performing singular value thresholding.
Default: 0.0001}

\item{max.iter}{The maximum number of iterations. Default: 100}

\item{verbose}{Whether return the information after each step of processing. Default: TRUE}
}
\description{
ENIGMA trace norm version
}
