% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/zzz.r
\name{error_handler}
\alias{error_handler}
\title{Error handler function for enigma_* functions}
\usage{
error_handler(x)
}
\arguments{
\item{x}{A response object from a call to \code{httr::GET}}
}
\description{
Error handler function for enigma_* functions
}
\keyword{internal}

