% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindCSE_DEG.R
\name{FindCSE_DEG}
\alias{FindCSE_DEG}
\title{Perform CTS-DEG analysis}
\usage{
FindCSE_DEG(
  object,
  y,
  FDR_control = TRUE,
  covariate = NULL,
  FoldChange = FALSE
)
}
\arguments{
\item{object}{ENIGMA object}

\item{FDR_control}{if use Wang et al., FDR controlled DEG analysis model. Default: TRUE}

\item{covariate}{The data.frame object contains the covariate information of each sample}

\item{FoldChange}{if output the FoldChange of each gene, if FALSE, then output the expression difference}
}
\value{
A list object contain the DEG results of each cell types
}
\description{
Analysis the differential expression genes for each cell type
}
\examples{
\dontrun{
DEG = FindCSE_DEG(object,y)
DEG = FindCSE_DEG(object,y,covariate=covariate)
head(DEG$celltype1)
}

}
