% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_proportion.R
\name{get_cell_proportion}
\alias{get_cell_proportion}
\title{Calculate the proportion of each cell types in bulk samples}
\usage{
get_cell_proportion(object, max.iter = 500)
}
\arguments{
\item{object}{ENIGMA object}

\item{max.iter}{the maximum iteration times for robust linear model. Default = 500}
}
\value{
the ENIGMA object contains the cell type fractions in object@result_cell_proportion
}
\description{
Calculate the proportion of each cell types in bulk samples
}
\examples{
\dontrun{
get_cell_proportion(egm)
}
}
