% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_proportion.R
\name{get_proportion}
\alias{get_proportion}
\title{Calculate the proportion of each cell types in bulk samples}
\usage{
get_proportion(X, ref)
}
\arguments{
\item{X}{Bulk gene expression dataset (gene*n) in linear space (non-log).}

\item{ref}{Cell-type specific gene expression profile (GEP, gene*prop) in linear space (non-log).}
}
\value{
A matrix with proportions of each cell types in bulk samples.
}
\description{
Calculate the proportion of each cell types in bulk samples
}
