% Generated by roxygen2 (4.0.2): do not edit by hand
\name{rate_limit}
\alias{rate_limit}
\title{Get rate limit data.}
\usage{
rate_limit(key = NULL, ...)
}
\arguments{
\item{key}{(character) Required. An Enigma API key. Supply in the function call, or store in
your \code{.Rprofile} file, or do \code{options(enigmaKey = "<your key>")}. Obtain an API key
by creating an account with Enigma at \url{https://enigma.io}, then obtain an API key from
your account page.}

\item{...}{Named options passed on to \code{\link[httr]{GET}}}
}
\description{
Get rate limit data.
}
\examples{
\donttest{
rate_limit()
}
}

