% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sce2array.R
\name{sce2array}
\alias{sce2array}
\title{Convert the output of ENIGMA CSE profiles into array format}
\usage{
sce2array(object, model_name = NULL, norm_output = TRUE)
}
\arguments{
\item{object}{the ENIGMA object}

\item{model_name}{name of the model, if null, the object@result_CSE or object@result_CSE_normalized would be used}

\item{norm_output}{Output the normalized CSE profiles}
}
\value{
an array with three dimensions.
}
\description{
Convert the ENIGMA CSE profile (SingleCellExperiments object) into array format (genes * samples * cell types)
}
\examples{
\dontrun{
egm = ENIGMA_l2max_norm(egm,model_tracker = TRUE, model_name = "Ken")
CSE = sce2array(egm, model_name = "Ken")
CSE = sce2array(egm)
}


}
