% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_correct.R
\name{batch_correct}
\alias{batch_correct}
\title{Remove batch effect with bulk and reference}
\usage{
batch_correct(
  object,
  varname_cell_type,
  n_pseudo_bulk = 100,
  sample_size = 1000,
  ncores = 6
)
}
\arguments{
\item{object}{ENIGMA object}

\item{varname_cell_type}{Varname in colData of reference to tell which column is cell type. Only avaliable when reference is single cell RNA-seq data.}

\item{n_pseudo_bulk}{Number of pseudo bulk samples generated from scRNA reference. Only avaliable when reference is single cell RNA-seq data. Default is 100.}

\item{sample_size}{Number of sampled cell for reconstituting pseudo-bulk samples. Default is 1000.}
}
\value{
For B-mode correction, ENIGMA object where object@bulk contains the corrected bulk gene expression profile. For S-mode correction, ENIGMA object where object@ref contains the corrected reference profile matrix.
}
\description{
This function would perform cross-platform batch effects correction. if object@ref_type = "single_cell", it would use S-mode batch correction method, if object@ref_type = "sort" or "aggre", it would use B-mode batch correction method. We note that this step is not essential step in deconvolution.
}
\examples{
\dontrun{
#S-mode
egm = batch_correct(egm, varname_cell_type = "main_celltype", n_pseudo_bulk=1000)
#B-mode
egm = batch_correct(egm)
}

}
