% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{enigma-package}
\alias{enigma}
\alias{enigma-package}
\title{enigma, an R client for Enigma.io}
\description{
Enigma holds government data and provides a really nice set of APIs for data, metadata, and stats on each of the datasets. That is, you can request a dataset itself, metadata on the dataset, and summary statistics on the columns of each dataset.
}
\details{
There are currently three functions

\itemize{
 \item enigma_data - Fetch and dataset, and filter on columns or rows.
 \item enigma_metadata - Get metadata on datasets.
 \item enigma_stats - Get columnwise statistics on datasets.
}
}

