% Generated by roxygen2 (4.0.0): do not edit by hand
\name{enigma_data}
\alias{enigma_data}
\title{Fetch a dataset from Enigma.}
\usage{
enigma_data(dataset = NULL, limit = 10, select = NULL, sort = NULL,
  page = NULL, where = NULL, search = NULL, key = NULL)
}
\arguments{
\item{dataset}{X}

\item{limit}{X}

\item{select}{X}

\item{sort}{X}

\item{page}{X}

\item{where}{X}

\item{search}{X}

\item{key}{X}
}
\description{
Fetch a dataset from Enigma.
}
\examples{
\dontrun{
enigma_data(dataset='us.gov.whitehouse.visitor-list')
}
}

