% Generated by roxygen2 (4.0.1): do not edit by hand
\name{enigma_fetch}
\alias{enigma_fetch}
\title{Get statistics on columns of a dataset from Enigma.}
\usage{
enigma_fetch(dataset = NULL, output = NULL, key = NULL, ...)
}
\arguments{
\item{dataset}{Dataset name. Required.}

\item{key}{(character) An Enigma API key. Supply in the function call, or store in your
\code{.Rprofile} file, or do \code{options(enigmaKey = "<your key>")}. Required.}

\item{...}{Named options passed on to \code{httr::GET}}
}
\description{
Get statistics on columns of a dataset from Enigma.
}
\examples{
\dontrun{
enigma_fetch(dataset='com.crunchbase.info.companies.acquisition')

library('httr')
enigma_fetch(dataset='com.crunchbase.info.companies.acquisition', config=verbose())
}
}

