% Generated by roxygen2 (4.0.2): do not edit by hand
\name{enigma_metadata}
\alias{enigma_metadata}
\title{Search for metadata on a dataset from Enigma.}
\usage{
enigma_metadata(dataset = NULL, key = NULL, ...)
}
\arguments{
\item{dataset}{Dataset name. Required.}

\item{key}{(character) An Enigma API key. Supply in the function call, or store in your
\code{.Rprofile} file, or do \code{options(enigmaKey = "<your key>")}}

\item{...}{Named options passed on to \code{httr::GET}}
}
\description{
Search for metadata on a dataset from Enigma.
}
\details{
Notice when you run the examples that the format of output is different for the
"parent nodes" vs. the "table nodes". Where the parent nodes have ouput$meta slots for
paths, immediate nodes and children tables, while the table nodes have ouput$meta slots for
info, table, ancestor datapaths, database boundary datapath, database boundary label, database
boundary tables, and paths, and an additional slot for description of table column attributes.
}
\examples{
\dontrun{
# Parent node response attributes
enigma_metadata(dataset='us.gov.whitehouse')
enigma_metadata(dataset='us.gov.dot.rita.trans-stats.air-carrier-statistics')

# Table node response attributes
## US white house visitor list
enigma_metadata(dataset='us.gov.whitehouse.visitor-list')

## Domestic Market Flight Statistics (Final Destination)
enigma_metadata('us.gov.dot.rita.trans-stats.air-carrier-statistics.t100d-market-all-carrier')
}
}

