% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_proportion.R
\name{get_cell_proportion}
\alias{get_cell_proportion}
\title{Calculate the proportion of each cell types in bulk samples}
\usage{
get_cell_proportion(object)
}
\arguments{
\item{object}{ENIGMA object}
}
\description{
Calculate the proportion of each cell types in bulk samples
}
