% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_proportion.R
\name{get_cell_proportion}
\alias{get_cell_proportion}
\title{Calculate the proportion of each cell types in bulk samples}
\usage{
get_cell_proportion(object, max.iter = 500, method = "RLR", nu.v)
}
\arguments{
\item{object}{ENIGMA object}

\item{max.iter}{the maximum iteration times for robust linear model. Default = 500}

\item{nu.v}{parameter needed for nu-regression, user could input a vector of nu, and CIBERSORT would automatically select the nu with the best fitness with users' data.}
}
\value{
the ENIGMA object contains the cell type fractions in object@result_cell_proportion
}
\description{
Calculate the proportion of each cell types in bulk samples
}
\examples{
\dontrun{
egm = get_cell_proportion(egm)
egm@result_cell_proportion
}
}
