% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rate_limit.R
\name{rate_limit}
\alias{rate_limit}
\title{Get rate limit data.}
\usage{
rate_limit(key = NULL, ...)
}
\arguments{
\item{key}{(character) Required. An Enigma API key. Supply in the function 
call, or store in your \code{.Rprofile} file, or do 
\code{options(enigmaKey = "<your key>")}. Obtain an API key by creating 
an account with Enigma at \url{http://enigma.io}, then obtain an API key 
from your account page.}

\item{...}{Named curl options passed on to \code{\link[crul]{HttpClient}}}
}
\description{
Get rate limit data.
}
\examples{
\dontrun{
rate_limit()
}
}

